+++ babel.config.js
@@ -3,20 +3,12 @@ module.exports = {
 		[
 			'@babel/env',
 			{
-				loose: true,
 				modules: 'auto',
+				corejs: 3,
 				useBuiltIns: 'usage',
 				targets: {
 					node: '12.20'
-					// browsers: [
-					// 	'> 1%',
-					// 	'last 2 versions',
-					// 	'not ie < 12',
-					// 	'not ie_mob < 12',
-					// 	'not op_mini all'
-					// ]
-				},
-				corejs: 3
+				}
 			}
 		],
 		'@babel/typescript'