+++ src/index.ts
@@ -1,4 +1,5 @@
 import { readFileSync } from 'fs'
+import { type MinifyOptions, minify } from 'terser'
 // import consola from 'consola'
 import type { Plugin } from 'rollup'
 
@@ -27,8 +28,39 @@ export interface Options {
 	 * @example ''
 	 */
 	// exportsModules: Record<string, string | string[]>
-	path: string
 	position?: 'before' | 'after'
+	/**
+	 * A string to prepend to the bundle
+	 */
+	intro?: string
+	/**
+	 * A string to append to the bundle
+	 */
+	outro?: string
+	/**
+	 * minify the codes
+	 */
+	minify?: boolean
+	/**
+	 * minify options for terser
+	 */
+	minifyOptions?: MinifyOptions
+}
+
+export interface OptionsPath extends Options {
+	/**
+	 * inject code path
+	 * @description Only one of “path” and “code” needs to be passed in, and "path" has higher priority than "code" when both are passed in
+	 */
+	path: string
+}
+
+export interface OptionsCode extends Options {
+	/**
+	 * inject code string
+	 * @description Only one of “path” and “code” needs to be passed in, and "path" has higher priority than "code" when both are passed in
+	 */
+	code: string
 }
 
 /**
@@ -37,20 +69,52 @@ export interface Options {
  * @param options - options
  * @returns Plugin - injectCode plugin
  */
-export default function injectCode(options: Options): Plugin {
-	if (!options.path) throw new Error('path is required')
+function injectCode(options: OptionsPath): Plugin
+function injectCode(options: OptionsCode): Plugin
+function injectCode(options: OptionsPath | OptionsCode): Plugin {
+	if ('path' in options && !options.path) throw new Error('"path" is required')
+	if (!('path' in options) && !('code' in options))
+		throw new Error('“path” and "code" must be passed into at least one')
+
 	return {
 		name: 'inject-code',
-		renderChunk(code) {
-			const resolvedUrl = require.resolve(options.path, {
-				paths: [process.cwd()]
-			})
-			// consola.info('[INJECT_CODE_INFO] resolvedUrl: ', resolvedUrl)
+		async renderChunk(code) {
+			let injectCode = '',
+				map
+			if ('path' in options) {
+				const resolvedUrl = require.resolve(options.path, {
+					paths: [process.cwd()]
+				})
+				// consola.info('[injectCode] resolvedUrl: ', resolvedUrl)
 
-			const INJECT_CODE = readFileSync(resolvedUrl, { encoding: 'utf-8' })
+				injectCode = readFileSync(resolvedUrl, { encoding: 'utf-8' })
+			} else if ('code' in options) {
+				injectCode = options.code
+			}
+
+			injectCode = `${'intro' in options ? options.intro + '\n' : ''}${injectCode}${
+				'outro' in options ? '\n' + options.outro : ''
+			}`
+
+			if (options.minify) {
+				const result = await minify(injectCode, options.minifyOptions)
+				injectCode = result.code || ''
+				map = typeof result.map === 'string' ? JSON.parse(result.map) : result.map
+
+				return {
+					code:
+						options.position === 'after'
+							? `${code.replace(/;*$/, ';')}\n${injectCode}`
+							: `${injectCode.replace(/;*$/, ';')}\n${code}`,
+					map
+				}
+			}
+
 			return options.position === 'after'
-				? `${code.replace(/;*$/, ';')}\n${INJECT_CODE}`
-				: `${INJECT_CODE.replace(/;*$/, ';')}\n${code}`
+				? `${code.replace(/;*$/, ';')}\n${injectCode}`
+				: `${injectCode.replace(/;*$/, ';')}\n${code}`
 		}
 	}
 }
+
+export default injectCode