+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Update logs
 
+## 2023.05.26 v1.1.0
+
+1. support new option: intro/outro/minify/minifyOptions/code
+2. Only one of “path” and “code” needs to be passed in, and "path" has higher priority than "code" when both are passed in
+3. upgrade all packages
+
 ## 2023.04.27 v1.0.0
 
 1. create project [README.md](./README.md)